[PHOENIX-7636]Moved Validator to different package and removed code duplicate.
