+++ egov/egov-tlweb/src/main/java/org/egov/tl/web/controller/transactions/closure/CreateLicenseClosureController.java
@@ -48,6 +48,7 @@
 package org.egov.tl.web.controller.transactions.closure;
 
 import org.egov.tl.entity.TradeLicense;
+import org.egov.tl.web.validator.closure.CreateLicenseClosureValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
